Work through some todos
No longer capturing output in test assertion when not asked to.
This generates purer/simpler test asserts for the most common use case.

Trailing whitespace is no longer significant.
You can't see it in a doc, so we don't assert for it in a test.

Incomplete REPL assertions no longer throw and now simply don't convert.
This does not include invalid Clojure, that'll still throw.