Address test order
Add test-ns-hook to generated tests to ensure test run order is
same a doc block order.

Document with special note on Kaocha which does not support test-ns-hook.
We are now naming tests so that they can be run in order in Kaocha,
which seems to run them in alphabetical order when no randomization is
requested.

Note that test-ns-hook prevents use of test fixtures.
That's ok for now, we aren't using fixtures yet and will address a solution
if we want/need specific test setup and teardown.

Also: fixed a booboo where I broke metadata generation.
Was due to shadowing a var.