Work-around odd rewrite-cljc behavior
I can't remember if I had a good reason to tag a map-qualifier node
as a :token in rewrite-cljc, but tokens are usually sexpr-able and
a map-qualifier is not.

Work-around for now by also checking if the node is not printable only
which is rewrite-clj geek for asking if it is sexpr-able.