A better description of clj-concordian
I wanted to make it clear that Concordion differs from Cucumber. I hope "far simpler" is not an overdrive. I do believe it is :) (I have also updated its readme to make it clearer to a newcomer what it does so that it is easier to understand without going to concordion.org)