Multiline eval expectations complete when parsable
We now know know that an eval expectation is complete when its form
parses successfully.

Added the lovely nubank matcher combinators dep to help with test
assertions.

Now that things are a bit more complicated, algorithm and naming in
code. It might be a bit easier to follow now? Hope so.