Fix multiline expectations (#35)
* Allows using leading comments in expected output for editor style

* Updated expected generated tests

* Fix typo in docstring

* Add unit test for removing leading comments

* Update expected generated tests

* Allow multiple leading comment characters

* Add test cases

* Add integration test case for leading semicolon support for adoc

* Add test cases based on more involved mixing of eval and stdout

* docs: examples: add section to explain multiline

Create a dedicated section to explain multiline output for example.md
and example.adoc.

These both explain to the user and act as tests for test-doc-blocks.

* Multiline eval expectations complete when parsable

We now know  that an eval expectation is complete when its form
parses successfully.

Added the lovely nubank matcher combinators dep to help with test
assertions.

Now that things are a bit more complicated, algorithm and naming in
code. It might be a bit easier to follow now? Hope so.

* review feedback: fix typo

* doc: examples: update single semicolon advice

We now would only ever need the single semicolon distinguisher for
`=stdout=>` and `=stderr=>` blocks.

* lint: add nubank/matcher-combinators config

* doc: changelog

Closes #33
---------

Co-authored-by: lread <lee@dlread.com>